requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
backoff==2.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langfuse==2.60.2
litellm==1.65.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
multidict==6.2.0
openai==1.69.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.3.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
pytest==8.3.5
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
scikit-learn==1.6.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.1
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0